requirements.txt
fastapi==0.76.0
pyserial==3.5
python-dotenv==0.20.0
uvicorn==0.17.6
